Player movement has been redone
input handler asset can now be removed.
Yes, i'm going through redoing basically everything.
