+++ Assets/Programming/Character/Input Handling and Movement.meta
@@ -1,0 +1,8 @@
+fileFormatVersion: 2
+guid: 124cd4c1f4e64064e901b793d4dbbb34
+folderAsset: yes
+DefaultImporter:
+  externalObjects: {}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
