+++ Assets/Programming/Character/Input Handling and Movement/PlayerControls.inputactions.meta
@@ -1,0 +1,10 @@
+fileFormatVersion: 2
+guid: b9977ed2e44091c43bd71efaabf09cdf
+ScriptedImporter:
+  internalIDToNameTable: []
+  externalObjects: {}
+  serializedVersion: 2
+  userData: 
+  assetBundleName: 
+  assetBundleVariant: 
+  script: {fileID: 11500000, guid: 8404be70184654265930450def6a9037, type: 3}
