+++ Assets/Programming/Character/Input Handling and Movement/PlayerMovementHandler.cs.meta
@@ -1,0 +1,11 @@
+fileFormatVersion: 2
+guid: c34fbf3c10d0ee64195612e035555a5e
+MonoImporter:
+  externalObjects: {}
+  serializedVersion: 2
+  defaultReferences: []
+  executionOrder: 0
+  icon: {instanceID: 0}
+  userData: 
+  assetBundleName: 
+  assetBundleVariant:
