--- Assets/Programming/Character/Input Handling/PlayerInputHandler.cs.meta
@@ -1,13 +1,0 @@
-fileFormatVersion: 2
-guid: 15c779fe09301c64aa4dbd3a95502f2b
-MonoImporter:
-  externalObjects: {}
-  serializedVersion: 2
-  defaultReferences:
-  - PlayerControls: {fileID: -944628639613478452, guid: b9977ed2e44091c43bd71efaabf09cdf,
-      type: 3}
-  executionOrder: 0
-  icon: {instanceID: 0}
-  userData: 
-  assetBundleName: 
-  assetBundleVariant:
